+++ src/main/java/com/snowhite/server/domain/entity/Card.java
@@ -1,6 +1,7 @@
-package com.snowhite.server.domain;
+package com.snowhite.server.domain.entity;
 
 import com.snowhite.server.domain.common.BaseEntity;
+import com.snowhite.server.domain.enums.CardType;
 import jakarta.persistence.*;
 
 @Entity
