+++ src/main/java/com/snowhite/server/domain/entity/File.java
@@ -1,4 +1,4 @@
-package com.snowhite.server.domain;
+package com.snowhite.server.domain.entity;
 
 import com.snowhite.server.domain.common.BaseEntity;
 import jakarta.persistence.*;
