+++ src/main/java/com/snowhite/server/domain/session/Room.java
@@ -1,5 +1,6 @@
-package com.snowhite.server.domain;
+package com.snowhite.server.domain.session;
 
+import com.snowhite.server.domain.entity.User;
 import lombok.Data;
 import lombok.NoArgsConstructor;
