+++ src/main/java/com/snowhite/server/security/filter/JwtWebFilter.java
@@ -1,7 +1,8 @@
-package com.snowhite.server.config;
+package com.snowhite.server.security.filter;
 
-import com.snowhite.server.domain.User;
-import com.snowhite.server.domain.UserRepository;
+import com.snowhite.server.security.jwt.JwtProvider;
+import com.snowhite.server.domain.entity.User;
+import com.snowhite.server.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
