+++ src/main/java/com/snowhite/server/web/dto/EmailDto.java
@@ -1,11 +1,10 @@
-package com.snowhite.server.dto;
+package com.snowhite.server.web.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
 @Getter
-public class LoginRequestDto {
+public class EmailDto {
     private String email;
-    private String password;
 }
