+++ src/main/java/com/snowhite/server/web/dto/LoginRequestDto.java
@@ -1,10 +1,11 @@
-package com.snowhite.server.dto;
+package com.snowhite.server.web.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
 @Getter
-public class EmailDto {
+public class LoginRequestDto {
     private String email;
+    private String password;
 }
